Fix: Revert space npm install and composer install to non-"optimized" variants
The promise was that alternative install commands would speed up the installation times, but measuring the actual CI automation times on Jetbrains taught me that the affects are either nil or even possibly negative.

So... don't use these optimizations until we understand better how to use them or why they're not speeding up as we had expected.